pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-automation-suite"
version = "1.0.0"
description = "Medical phrase and snippet automation system"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Medical Automation Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Topic :: Office/Business",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "Flask>=2.3.0",
    "flask-cors>=4.0.0",
    "requests>=2.31.0",
    "pynput>=1.7.6",
    "pyperclip>=1.8.2",
    "pyautogui>=0.9.54",
    "pywebview>=4.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
    "pylint>=2.17.0",
    "safety>=2.3.0",
    "bandit>=1.7.5",
]

[project.scripts]
medical-server = "medical_automation.server:main"
medical-client = "medical_automation.client.medical_client:main"
snippet-expander = "medical_automation.client.snippet_expander:main"
snippet-manager = "medical_automation.client.snippet_manager:main"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/medical_automation --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pylint.messages_control]
max-line-length = 88
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
]

[tool.bandit]
exclude_dirs = ["tests", "venv"]
tests = ["B201", "B301"]
skips = ["B101", "B601"]